+++ packages/db/src/index.ts
@@ -2,6 +2,7 @@ export * from './schema';
 export * from './client';
 export * from './migrate';
 export * from './notes-repository';
+export * from './note-versions-repository';
 export * from './spaces-repository';
 export * from './search-repository';
 export * from './tokens-repository';